+++ packages/contracts/src/index.ts
@@ -6,6 +6,7 @@ export * from "./programme";
 export * from "./reporting";
 export * from "./experience";
 export * from "./customer-export";
+export * from "./entitlements";
 
 export const commerceEnvelopeV1 = z.object({
   version: z.literal("1"),